+++ src/index.html
@@ -12,18 +12,18 @@
   <body>
     <div class="container">
       <div class="title">
-        <h4>Today's ToDo</h4>
-        <i class="uil uil-redo"></i>
+        <h4>Bheki's ToDo</h4>
+
       </div>
       <hr />
       <form>
-        <input type="text" id="new-task" placeholder="Add to your list..." />
+        <input type="text" id="new-task" placeholder="Add to your list...😊" />
         <button type="submit"><i class="uil uil-enter"></i></button>
       </form>
       <ul id="task-list"></ul>
 
       <hr />
-      <a href="#">Clear all completed</a>
+      <a id="clear-all" href="#">Clear all completed</a>
     </div>
   </body>
 </html>
